requirements.txt
tqdm
loguru
flax
dm-haiku
optax
chex
transformers
tokenizers
datasets[jax]
autoflake
yapf
mkdocstrings[crystal,python]
mkdocs-material
rich
orbax-checkpoint
portpicker
wandb
sentencepiece
jax-smi